view api 연동 완료 및 예외처리 및 테스트 코드 작성

+++ src/main/java/com/future/practice/domain/board/controller/CommentController.java
@@ -6,7 +6,10 @@ import com.future.practice.domain.board.service.CommentService;
 import com.future.practice.global.constant.ResponseMessage;
 import com.future.practice.global.dto.ResponseDefaultDto;
 import com.future.practice.global.entity.User;
+import com.future.practice.global.exception.custom.ServerErrorException;
+import com.future.practice.global.exception.custom.UserNotExistException;
 import lombok.RequiredArgsConstructor;
+import lombok.extern.slf4j.Slf4j;
 import org.springframework.http.HttpHeaders;
 import org.springframework.http.ResponseEntity;
 import org.springframework.web.bind.annotation.*;
@@ -14,6 +17,7 @@ import org.springframework.web.bind.annotation.*;
 import javax.servlet.http.HttpSession;
 
 @RequestMapping("/v1")
+@Slf4j
 @RequiredArgsConstructor
 @RestController("")
 public class CommentController {
@@ -22,49 +26,72 @@ public class CommentController {
 
     @PostMapping("/board/{board_seq}/comment")
     public ResponseEntity<ResponseDefaultDto> insertComment(@PathVariable("board_seq") int board_seq,
-                                                            @RequestBody CommentDto commentDto,
+                                                            CommentDto commentDto,
                                                                         HttpSession session){
-        User user = (User) session.getAttribute("user");
-        commentService.insertComment(commentDto, board_seq, user);
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_COMMENT_INSERT_MESSAGE).build());
+        try {
+            if (session == null) throw new UserNotExistException();
+            User user = (User) session.getAttribute("user");
+            commentService.insertComment(commentDto, board_seq, user);
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_COMMENT_INSERT_MESSAGE).build());
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ServerErrorException();
+        }
     }
 
     @PostMapping("/comment/{comment_seq}")
     public ResponseEntity<ResponseDefaultDto>  insertBigComment(@PathVariable("comment_seq") int comment_seq,
-                                                                @RequestBody CommentDto commentDto,
+                                                                CommentDto commentDto,
                                                                 HttpSession session){
-        User user = (User) session.getAttribute("user");
-        commentService.insertBigComment(commentDto, comment_seq, user);
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_BIG_COMMENT_INSERT_MESSAGE).build());
+        try {
+            if (session == null) throw new UserNotExistException();
+            User user = (User) session.getAttribute("user");
+            commentService.insertBigComment(commentDto, comment_seq, user);
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_BIG_COMMENT_INSERT_MESSAGE).build());
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ServerErrorException();
+        }
     }
 
     @DeleteMapping("/comment/{comment_seq}")
     public ResponseEntity<ResponseDefaultDto>  deleteComment(@PathVariable("comment_seq") int comment_seq
                                                             ,HttpSession session){
-        User user = (User)session.getAttribute("user");
-        commentService.deleteComment(comment_seq, user);
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_COMMENT_DELETE_MESSAGE).build());
+        try {
+            if (session == null) throw new UserNotExistException();
+            User user = (User) session.getAttribute("user");
+            commentService.deleteComment(comment_seq, user);
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_COMMENT_DELETE_MESSAGE).build());
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ServerErrorException();
+        }
     }
 
     @DeleteMapping("/bigComment/{big_comment_seq}")
     public ResponseEntity<ResponseDefaultDto>  deleteBigComment(@PathVariable("big_comment_seq") int big_comment_seq
                                                                 ,HttpSession session){
-
-        User user = (User)session.getAttribute("user");
-        commentService.deleteBigComment(big_comment_seq, user);
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_BIG_COMMENT_DELETE_MESSAGE).build());
+        try {
+            if (session == null) throw new UserNotExistException();
+            User user = (User) session.getAttribute("user");
+            commentService.deleteBigComment(big_comment_seq, user);
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_BIG_COMMENT_DELETE_MESSAGE).build());
+        }catch (Exception e){
+            e.printStackTrace();
+            throw new ServerErrorException();
+        }
     }
 
 }
