+++ src/main/java/com/future/practice/domain/board/service/BoardServiceImpl.java
@@ -10,6 +10,8 @@ import com.future.practice.global.constant.Common;
 import com.future.practice.global.entity.Board;
 import com.future.practice.global.entity.Comment;
 import com.future.practice.global.entity.User;
+import com.future.practice.global.exception.ErrorCode;
+import com.future.practice.global.exception.custom.*;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.stereotype.Service;
@@ -26,16 +28,22 @@ public class BoardServiceImpl implements BoardService {
     private final CommentMapper commentMapper;
     @Override
     public void insertBoardService(BoardDto boardDto, User user ) {
+        if(boardDto.getTitle().equals("")) throw new BoardTitleNotExistException(); // 게시물 제목 없을 시
+        if(boardDto.getContent().equals("")) throw new BoardContentNotExistException(); // 게시물 내용 없을 시
         boardMapper.save(boardDto.toEntity(user.getUserEmail()));
     }
     @Override
     public void updateBoardService(BoardDto boardDto, long boardSeq, User user) {
+        if(boardDto.getTitle().equals("")) throw new BoardTitleNotExistException(); // 게시물 제목 없을 시
+        if(boardDto.getContent().equals("")) throw new BoardContentNotExistException(); // 게시물 내용 없을 시
         boardMapper.updateByBoardSeq(boardDto.toEntity(boardSeq));
     }
 
     @Override
     public void deleteBoardService(long boardSeq, User user) {
         Board board = Board.builder().boardSeq(boardSeq).boardUserEmail(user.getUserEmail()).build();
+        if(boardMapper.findOneByBoardSeq(boardSeq)==null ) throw new BoardNotFoundException(); // 게시물 존재 X
+        if(boardMapper.findOneByBoardSeqAndBoardUserEmail(board)==null) throw new BoardNotAccessException(); // 게시물 권한 확인
         boardMapper.deleteByBoardSeq(board);
     }
 
@@ -47,8 +55,10 @@ public class BoardServiceImpl implements BoardService {
         if(map.get("page") ==null) map.put("page",1);
 
         map.put("page", (Integer.parseInt(map.get("page").toString())-1 )* Common.BOARD_NUM);
-        log.info("board length : " + boardMapper.findAllLength(""));
+        List<Board> board = boardMapper.findAll(map);
+        int page = (int)map.get("page");
 
+        if(board.size()==0) throw new BoardNotFoundException(); // 게시물이 존재하지 않음
         return ResponseBoardDto.builder()
                 .board(boardMapper.findAll(map))
                 .page((int)map.get("page")) // Integer.parseInt(map.get("page").toString)
@@ -59,8 +69,8 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public ResponseBoardDetailDto selectBoardDetailService(long boardSeq) {
 
-        log.info(boardMapper.findOneByBoardSeq(boardSeq).toString());
-
+        Board board = boardMapper.findOneByBoardSeq(boardSeq);
+        if(board==null) throw new BoardNotFoundException();
         List<CommentAll> commentAll = new ArrayList<>();
         List<Comment> comment = commentMapper.findAllByCommentBoardSeq(boardSeq);
         for(Comment commentList : comment){
@@ -80,6 +90,7 @@ public class BoardServiceImpl implements BoardService {
     @Override
     public ResponseBoardDto selectSearchBoardService(Map<String, Object> map) {
         map.put("board_num", Common.BOARD_NUM);
+        if(map.get("page") ==null) map.put("page",1);
         map.put("page", (Integer.parseInt(map.get("page").toString())-1 )* Common.BOARD_NUM);
         return ResponseBoardDto
                 .builder()
