+++ src/main/java/com/future/practice/domain/user/controller/UserController.java
@@ -5,6 +5,10 @@ import com.future.practice.domain.user.service.UserService;
 import com.future.practice.global.constant.ResponseMessage;
 import com.future.practice.global.dto.ResponseDefaultDto;
 import com.future.practice.global.entity.User;
+import com.future.practice.global.exception.ErrorCode;
+import com.future.practice.global.exception.custom.CustomException;
+import com.future.practice.global.exception.custom.ServerErrorException;
+import com.future.practice.global.exception.custom.UserAlreadyExistException;
 import lombok.RequiredArgsConstructor;
 import lombok.extern.slf4j.Slf4j;
 import org.springframework.beans.factory.annotation.Autowired;
@@ -27,21 +31,16 @@ public class UserController {
 
     @PostMapping("/login")
     public ResponseEntity<ResponseDefaultDto> login(HttpSession session, UserDto.Login loginDto){
-        User user = userService.loginService(loginDto);
-        session.setAttribute("user",user); //session
-        log.info("loginDto email : "+ loginDto.getEmail() +" password :" + loginDto.getPassword());
-        //token 생성, Session 생성
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_LOGIN_MESSAGE).build());
-    }
+        if(session.getAttribute("user")!=null)  throw new UserAlreadyExistException();
+            User user = userService.loginService(loginDto);
+            session.setAttribute("user", user); //session
+            log.info("loginDto email : " + loginDto.getEmail() + " password :" + loginDto.getPassword());
+            //token 생성, Session 생성
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_LOGIN_MESSAGE).build());
 
-    @GetMapping("/logout")
-    public ResponseEntity<ResponseDefaultDto> logout(){
-
-        log.info("logout");
-        return null;
     }
 
     @PostMapping("/sign")
@@ -52,17 +51,19 @@ public class UserController {
         return ResponseEntity.ok()
                 .headers(headers)
                 .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_SIGN_MESSAGE).build());
+
     }
 
     @PutMapping("")
     public ResponseEntity<ResponseDefaultDto> updateUser( UserDto.Inform informDto, HttpSession session){
-        log.info("update User");
-        log.info("inform : password : " + informDto.getPassword() + "name" + informDto.getName() + "phone" + informDto.getPhone());
-        userService.updateService(informDto, ((User)session.getAttribute("user")).getUserEmail());
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_USER_UPDATE_MESSAGE).build());
+            log.info("update User");
+            if (session == null) throw new UserAlreadyExistException();
+
+            userService.updateService(informDto, ((User) session.getAttribute("user")).getUserEmail());
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_USER_UPDATE_MESSAGE).build());
     }
     @GetMapping("")
     public ResponseEntity<User> selectUser(HttpSession session){
@@ -76,12 +77,16 @@ public class UserController {
 
     @DeleteMapping("")
     public ResponseEntity<ResponseDefaultDto> deleteUser(HttpSession session){
-        log.info("delete User");
-        userService.deleteService(((User)session.getAttribute("user")).getUserEmail());
+        try {
+            if(session==null) throw new UserAlreadyExistException();
+            userService.deleteService(((User) session.getAttribute("user")).getUserEmail());
 
-        HttpHeaders headers = new HttpHeaders();
-        return ResponseEntity.ok()
-                .headers(headers)
-                .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_USER_DELETE_MESSAGE).build());
+            HttpHeaders headers = new HttpHeaders();
+            return ResponseEntity.ok()
+                    .headers(headers)
+                    .body(ResponseDefaultDto.builder().code(200).message(ResponseMessage.RESPONSE_USER_DELETE_MESSAGE).build());
+        }catch (Exception e){
+            throw new ServerErrorException();
+        }
     }
 }
